# api/candidate/openapi.yaml
openapi: 3.1.0
info: { title: FSP Platform — Candidate Service, version: 1.0.0 }
servers: [{ url: http://localhost:8083 }]
security: [{ bearerAuth: [] }]
paths:
...
  /me/profile:
    get:
      summary: Профиль текущего соискателя
      operationId: getMyProfile
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "200":
          description: Профиль кандидата
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CandidateProfile" }
    patch:
      summary: Обновить профиль соискателя
      operationId: updateMyProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProfileRequest" }
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "200":
          description: Обновлённый профиль
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CandidateProfile" }
  /me/contacts:
    get:
      summary: Контакты текущего соискателя
      operationId: getMyContacts
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "200":
          description: Контакты
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CandidateContacts" }
    put:
      summary: Заменить контакты соискателя
      operationId: replaceMyContacts
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CandidateContacts" }
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "200":
          description: Обновлено
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CandidateContacts" }
  /me/visibility:
    get:
      summary: Настройки видимости профиля
      operationId: getMyVisibility
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "200":
          description: Настройки видимости
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Visibility" }
    put:
      summary: Обновить настройки видимости
      operationId: updateMyVisibility
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Visibility" }
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "200": { description: Обновлено }
  /me/resumes:
    get:
      summary: Список резюме
      operationId: listResumes
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "200":
          description: Резюме
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Resume" }
    post:
      summary: Создать резюме вручную
      operationId: createResume
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResumeInput" }
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "201":
          description: Создано
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Resume" }
  /me/resumes/upload:
    post:
      summary: Загрузить PDF-резюме
      operationId: uploadResumePDF
      description: Загрузка PDF с автораспознаванием
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
              required: [file]
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "201":
          description: Резюме создано, поле parsed содержит извлечённые данные
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Resume" }
  /me/experiences:
    get:
      summary: Опыт работы
      operationId: listExperiences
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "200":
          description: Опыт работы
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Experience" }
    post:
      summary: Добавить место работы
      operationId: addExperience
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExperienceInput" }
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "201":
          description: Добавлено
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Experience" }
  /me/experiences/{experienceId}:
    delete:
      summary: Удалить место работы
      operationId: deleteExperience
      parameters:
        - name: experienceId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "204": { description: Удалено }
  /me/technologies:
    get:
      summary: Стек текущего соискателя
      operationId: listMyTechnologies
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "200":
          description: Выбранные технологии
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CandidateTechnology" }
    put:
      summary: Заменить стек технологий
      operationId: replaceTechnologies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: "#/components/schemas/CandidateTechnology" }
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "200": { description: Обновлено }
  /me/category:
    get:
      summary: Текущая категория и грейд
      operationId: getMyCategory
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "200":
          description: Текущая категория/грейд и история
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryState" }
  /me/fsp:
    get:
      summary: Привязка ФСП и достижения
      operationId: getMyFSP
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "200":
          description: Привязка к ФСП ID и достижения (может быть пустым)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FSPState" }
    put:
      summary: Привязать идентификатор участника ФСП
      operationId: linkFSP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { fspMemberId: { type: string } }
              required: [fspMemberId]
      responses:
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "200": { description: Привязано }

  # работодательский доступ (использует internal-токен)
  /candidates/{userId}:
    get:
      summary: Карточка кандидата для работодателя
      operationId: getCandidateForEmployer
      description: |
        Возвращает карточку кандидата. Контакты маскируются, если нет
        accepted-приглашения или отклика между этим работодателем и кандидатом.
        Раскрытие определяется сервером по contact_reveals; query-параметров для этого нет.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Карточка кандидата
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CandidateCard" }
        "403": { description: Только для работодателей }
        "404": { description: Не найдено }
  /internal/reveals:
    post:
      summary: Зафиксировать раскрытие контактов
      operationId: internalRecordReveal
      description: |
        Вызывается interaction-сервисом при accepted-приглашении или
        самостоятельном отклике кандидата. После этого работодатель увидит контакты.
      parameters:
        - in: header
          name: X-Internal-Token
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                candidateUserId: { type: string, format: uuid }
                employerUserId: { type: string, format: uuid }
                entityType: { type: string, enum: [invitation, application] }
                entityId: { type: string, format: uuid }
                reason: { type: string, enum: [invitation_accepted, application_sent] }
              required: [candidateUserId, employerUserId, entityType, entityId, reason]
      responses:
        "204": { description: OK }
        "403": { description: Неверный internal-токен }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Доступ запрещён
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        code: { type: string, example: BAD_REQUEST }
        message: { type: string }
      required: [code, message]
    CandidateProfile:
      type: object
      properties:
        userId: { type: string, format: uuid }
        firstName: { type: [string, "null"] }
        lastName: { type: [string, "null"] }
        middleName: { type: [string, "null"] }
        headline: { type: [string, "null"] }
        about: { type: [string, "null"] }
        location: { type: [string, "null"] }
        yearsExperience: { type: [number, "null"], format: float }
        categoryId: { type: [string, "null"], format: uuid }
        gradeId: { type: [string, "null"], format: uuid }
        specializationId: { type: [string, "null"], format: uuid }
        fspMemberId: { type: [string, "null"] }
        salaryMin: { type: [integer, "null"], minimum: 0 }
        salaryMax: { type: [integer, "null"], minimum: 0 }
        salaryCurrency: { type: [string, "null"], default: RUB }
        softSkills:
          type: array
          items: { type: string }
          maxItems: 20
        updatedAt: { type: string, format: date-time }
      required: [userId, updatedAt]
    UpdateProfileRequest:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        middleName: { type: string }
        headline: { type: string }
        about: { type: string }
        location: { type: string }
        yearsExperience: { type: number, minimum: 0 }
        salaryMin: { type: integer, minimum: 0 }
        salaryMax: { type: integer, minimum: 0 }
        salaryCurrency: { type: string }
        specializationId: { type: string, format: uuid }
        softSkills:
          type: array
          items: { type: string }
          maxItems: 20
    CandidateContacts:
      type: object
      properties:
        email: { type: [string, "null"], format: email }
        phone: { type: [string, "null"] }
        telegram: { type: [string, "null"] }
        github: { type: [string, "null"] }
        linkedin: { type: [string, "null"] }
        website: { type: [string, "null"], format: uri }
        masked: { type: boolean }
        maskedFields: { type: array, items: { type: string } }
    Visibility:
      type: object
      properties:
        contacts: { type: boolean }
        links: { type: boolean }
        fsp: { type: boolean }
        experience: { type: boolean }
        resume: { type: boolean }
        salary: { type: boolean }
        softSkills: { type: boolean }
    Resume:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        source: { type: string, enum: [manual, pdf_upload, generated] }
        isPrimary: { type: boolean }
        parsed: { type: [object, "null"], additionalProperties: true }
        createdAt: { type: string, format: date-time }
    ResumeInput:
      type: object
      properties:
        title: { type: string }
        content: { type: string }
        isPrimary: { type: boolean, default: false }
      required: [title, content]
    Experience:
      type: object
      properties:
        id: { type: string, format: uuid }
        company: { type: string }
        position: { type: string }
        startedAt: { type: string, format: date }
        endedAt: { type: [string, "null"], format: date }
        description: { type: [string, "null"] }
      required: [id, company, position, startedAt]
    ExperienceInput:
      type: object
      properties:
        company: { type: string }
        position: { type: string }
        startedAt: { type: string, format: date }
        endedAt: { type: [string, "null"], format: date }
        description: { type: string }
      required: [company, position, startedAt]
    CandidateTechnology:
      type: object
      properties:
        technologyId: { type: string, format: uuid }
        level: { type: integer, minimum: 1, maximum: 5 }
      required: [technologyId]
    CategoryState:
      type: object
      properties:
        categoryId: { type: [string, "null"], format: uuid }
        gradeId: { type: [string, "null"], format: uuid }
        specializationId: { type: [string, "null"], format: uuid }
        history:
          type: array
          items:
            type: object
            properties:
              categoryId: { type: string, format: uuid }
              gradeId: { type: string, format: uuid }
              specializationId: { type: string, format: uuid }
              reason: { type: string, enum: [initial_test, retake, manual] }
              effectiveFrom: { type: string, format: date-time }
              effectiveTo: { type: [string, "null"], format: date-time }
    FSPState:
      type: object
      properties:
        fspMemberId: { type: [string, "null"] }
        linkedAt: { type: [string, "null"], format: date-time }
        achievements:
          type: array
          items: { $ref: "#/components/schemas/FSPAchievement" }
    FSPAchievement:
      type: object
      properties:
        id: { type: string, format: uuid }
        externalId: { type: [string, "null"] }
        eventName: { type: string }
        eventDate: { type: [string, "null"], format: date }
        place: { type: [integer, "null"] }
        category: { type: [string, "null"] }
        score: { type: [number, "null"] }
        weight: { type: integer }
    CandidateCard:
      type: object
      properties:
        userId: { type: string, format: uuid }
        displayName: { type: string }
        headline: { type: [string, "null"] }
        categoryId: { type: [string, "null"], format: uuid }
        gradeId: { type: [string, "null"], format: uuid }
        specializationId: { type: [string, "null"], format: uuid }
        yearsExperience: { type: [number, "null"] }
        location: { type: [string, "null"] }
        softSkills: { type: array, items: { type: string } }
        salary: { $ref: "#/components/schemas/CandidateCardSalary" }
        technologies:
          type: array
          items: { $ref: "#/components/schemas/CandidateTechnology" }
        fsp: { $ref: "#/components/schemas/CandidateCardFsp" }
        contacts: { $ref: "#/components/schemas/CandidateContacts" }
    CandidateCardSalary:
      type: object
      properties:
        min: { type: [integer, "null"] }
        max: { type: [integer, "null"] }
        currency: { type: [string, "null"] }
        masked: { type: boolean }
    CandidateCardFsp:
      type: object
      properties:
        hasFSP: { type: boolean }
        achievementsCount: { type: integer }
        bestPlace: { type: [integer, "null"] }